refactor: merge test and schema inspection functions into db.py
